Move metering to package
